Fixed an oversight in the utilization calculation that broke SDF + task, for good.

+++ java-common/src/main/java/idesyde/common/TiledMultiCore.java
@@ -12,25 +12,25 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public record TiledMultiCore(
-    @JsonProperty("communication_elements_bit_per_sec_per_channel") Map<String, Double> communicationElementsBitPerSecPerChannel,
-    @JsonProperty("communication_elements_max_channels") Map<String, Long> communicationElementsMaxChannels,
-    @JsonProperty("interconnect_topology_dsts") List<String> interconnectTopologyDsts,
-    @JsonProperty("interconnect_topology_srcs") List<String> interconnectTopologySrcs,
-    Set<String> memories,
-    @JsonProperty("network_interfaces") Set<String> networkInterfaces,
-    @JsonProperty("pre_computed_paths") Map<String, Map<String, List<String>>> preComputedPaths,
-    Set<String> processors,
-    @JsonProperty("processors_frequency") Map<String, Long> processorsFrequency,
-    @JsonProperty("processors_provisions") Map<String, Map<String, Map<String, Double>>> processorsProvisions,
-    Set<String> routers,
-    @JsonProperty("tile_memory_sizes") Map<String, Long> tileMemorySizes
-) implements DecisionModel {
+        @JsonProperty("communication_elements_bit_per_sec_per_channel") Map<String, Double> communicationElementsBitPerSecPerChannel,
+        @JsonProperty("communication_elements_max_channels") Map<String, Integer> communicationElementsMaxChannels,
+        @JsonProperty("interconnect_topology_dsts") List<String> interconnectTopologyDsts,
+        @JsonProperty("interconnect_topology_srcs") List<String> interconnectTopologySrcs,
+        Set<String> memories,
+        @JsonProperty("network_interfaces") Set<String> networkInterfaces,
+        @JsonProperty("pre_computed_paths") Map<String, Map<String, List<String>>> preComputedPaths,
+        Set<String> processors,
+        @JsonProperty("processors_frequency") Map<String, Long> processorsFrequency,
+        @JsonProperty("processors_provisions") Map<String, Map<String, Map<String, Double>>> processorsProvisions,
+        Set<String> routers,
+        @JsonProperty("tile_memory_sizes") Map<String, Long> tileMemorySizes) implements DecisionModel {
     @Override
     public DecisionModelHeader header() {
         return new DecisionModelHeader(
                 "TiledMultiCore",
-                Stream.concat(memories.stream(), Stream.concat(processors.stream(), Stream.concat(routers.stream(), networkInterfaces.stream()))).collect(Collectors.toSet()),
-                Optional.empty()
-        );
+                Stream.concat(memories.stream(),
+                        Stream.concat(processors.stream(), Stream.concat(routers.stream(), networkInterfaces.stream())))
+                        .collect(Collectors.toSet()),
+                Optional.empty());
     }
 }
